Organize adaptor exports
