+++ packages/tendermint-rpc/src/adaptors/v0-33/hasher.spec.ts
@@ -1,8 +1,8 @@
 import { fromBase64, fromHex } from "@cosmjs/encoding";
 import { ReadonlyDate } from "readonly-date";
 
-import { ReadonlyDateWithNanoseconds } from "../responses";
-import { TxBytes } from "../types";
+import { ReadonlyDateWithNanoseconds } from "../../responses";
+import { TxBytes } from "../../types";
 import { hashBlock, hashTx } from "./hasher";
 
 describe("Hasher", () => {
