+++ packages/tendermint-rpc/types/adaptors/v0-33/hasher.d.ts
@@ -1,4 +1,4 @@
-import { Header } from "../responses";
-import { BlockHash, TxBytes, TxHash } from "../types";
+import { Header } from "../../responses";
+import { BlockHash, TxBytes, TxHash } from "../../types";
 export declare function hashTx(tx: TxBytes): TxHash;
 export declare function hashBlock(header: Header): BlockHash;
