+++ packages/tendermint-rpc/types/adaptors/v0-33/hasher.spec.d.ts
@@ -1,0 +1,1 @@
+export {};
