--- packages/tendermint-rpc/types/v0-33/index.d.ts
@@ -1,2 +1,0 @@
-import { Adaptor } from "../adaptor";
-export declare const v0_33: Adaptor;
